requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
discord.py==1.7.3
frozenlist==1.3.0
idna==3.3
multidict==5.2.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.19.2
pytube @ git+https://github.com/glubsy/pytube@5fc27b72180f6bb76d43633c29a89eff5df1f70e
six==1.16.0
typing_extensions==4.0.1
yarl==1.7.2
